requirements.txt
fastapi
uvicorn
sqlalchemy
databases
python-jose[cryptography]
passlib[bcrypt]
pyotp
qrcode[pil] 